+++ src/api/v1/app/urls.py
@@ -11,22 +11,28 @@ from . import views
 
 urlpatterns = [
     # -------------------------------------------------------------------------
-    # --- Attachments
+    # --- Attachments.
+    # -------------------------------------------------------------------------
     re_path(r"^tmp-upload/$", views.tmp_upload, name="api-tmp-upload"),
+
     re_path(r"^upload/(?P<upload_type>[\w_-]+)/(?P<upload_id>[\w_-]+)/$", views.upload_details, name="api-upload-details"),
-    re_path(r"^link/(?P<link_type>[\w_-]+)/(?P<link_id>[\w_-]+)/$", views.link_details, name="api-link-details"),
+    re_path(r"^upload/(?P<upload_type>[\w_-]+)/(?P<upload_id>[\w_-]+)/private/$", views.upload_private, name="api-upload-private"),
+    re_path(r"^upload/(?P<upload_type>[\w_-]+)/(?P<upload_id>[\w_-]+)/hidden/$", views.upload_hidden, name="api-upload-hidden"),
 
     # -------------------------------------------------------------------------
     # --- Comments.
+    # -------------------------------------------------------------------------
     re_path(r"^comments/$", views.comment_list, name="api-comment-list"),
     re_path(r"^comments/(?P<comment_id>[\w_-]+)/$", views.comment_details, name="api-comment-details"),
 
     # -------------------------------------------------------------------------
     # --- Complaints.
+    # -------------------------------------------------------------------------
     re_path(r"^complaints/$", views.complaint_list, name="api-complaint-list"),
 
     # -------------------------------------------------------------------------
     # --- Ratings.
+    # -------------------------------------------------------------------------
     re_path(r"^ratings/$", views.rating_list, name="api-rating-list"),
     re_path(r"^ratings/(?P<rating_id>[\w_-]+)/$", views.rating_details, name="api-rating-details"),
 ]
