+++ src/api/v1/app/views/__init__.py
@@ -1,0 +1,10 @@
+"""
+(C) 2013-2025 Copycat Software, LLC. All Rights Reserved.
+"""
+
+# pylint: disable=wildcard-import
+# pylint: disable=unused-wildcard-import
+from api.v1.app.views.attachments import *
+from api.v1.app.views.comments import *
+from api.v1.app.views.complaints import *
+from api.v1.app.views.rating import *
