+++ src/api/v1/app/views/rating.py
@@ -1,0 +1,323 @@
+"""
+(C) 2013-2025 Copycat Software, LLC. All Rights Reserved.
+"""
+
+import inspect
+import logging
+
+from django.conf import settings
+from django.contrib.auth.models import User
+from django.contrib.contenttypes.models import ContentType
+from django.template import loader
+from django.utils.translation import ugettext_lazy as _
+
+from rest_framework import (
+    status,
+    views,
+    viewsets,
+    parsers,
+    renderers,
+    mixins)
+from rest_framework.permissions import (
+    AllowAny,
+    IsAdminUser,
+    IsAuthenticated,
+    IsAuthenticatedOrReadOnly)
+from rest_framework.renderers import JSONRenderer
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+from annoying.functions import get_object_or_None
+from termcolor import cprint
+
+from ddcore.models import (
+    AttachedDocument,
+    AttachedImage,
+    AttachedUrl,
+    AttachedVideoUrl,
+    Comment,
+    Complaint,
+    Rating,
+    TemporaryFile)
+
+# pylint: disable=import-error
+from accounts.models import UserProfile
+from api.auth import CsrfExemptSessionAuthentication
+# from api.v1.events.utils import (
+#     event_access_check_required,
+#     event_org_staff_member_required)
+# from api.v1.organizations.utils import (
+#     organization_access_check_required,
+#     organization_staff_member_required)
+from app import logconst
+from app.decorators import log_default
+from app.logformat import Format
+from blog.models import Post
+from events.models import Event
+from organizations.models import Organization
+
+
+logger = logging.getLogger(__name__)
+
+
+# =============================================================================
+# ===
+# === RATINGS
+# ===
+# =============================================================================
+class RatingListViewSet(APIView):
+    """Rating List View Set."""
+
+    # authentication_classes = (CsrfExemptSessionAuthentication, )
+    permission_classes = (IsAuthenticated, )
+    renderer_classes = (JSONRenderer, )
+    # serializer_class = RatingSerializer
+    # model = Rating
+
+    @log_default(my_logger=logger)
+    def post(self, request):
+        """POST: Rating create.
+
+        Parameters
+        ----------
+        account_id          :int
+        event_id            :int
+        organization_id     :int
+        place_id            :int
+        post_id             :int
+
+        author_rating       :int
+        rating              :int
+
+        review_text         :str
+
+                Example Payload:
+
+                    {
+                        "event_id":     1,
+                        "rating":       5,
+                        "review_text":  "Event Review Text"
+                    }
+
+        Returns
+        -------
+                            :dict
+
+                Example Payload:
+
+                    {
+                        "message":      "Successfully added the Complaint.",
+                    }
+
+        Raises
+        ------
+
+        """
+
+        # ---------------------------------------------------------------------
+        # --- Retrieve Data from the Request.
+        # ---------------------------------------------------------------------
+        account_id = request.data.get("account_id", "")
+        event_id = request.data.get("event_id", "")
+        organization_id = request.data.get("organization_id", "")
+        place_id = request.data.get("place_id", "")
+        post_id = request.data.get("post_id", "")
+        author_rating = request.data.get("author_rating", "")
+        rating = request.data.get("rating", "")
+        review_text = request.data.get("review_text", "")
+
+        cprint(f"[---  DUMP   ---] ACCOUNT      ID : {account_id}\n"
+               f"                  EVENT        ID : {event_id}\n"
+               f"                  ORGANIZATION ID : {organization_id}\n"
+               f"                  PLACE        ID : {place_id}\n"
+               f"                  POST         ID : {post_id}\n"
+               f"                  AUTHOR   RATING : {author_rating}\n"
+               f"                           RATING : {rating}\n"
+               f"                  REVIEW     TEXT : {review_text}", "yellow")
+
+        # ---------------------------------------------------------------------
+        # --- Verify Request.
+        # ---------------------------------------------------------------------
+        if (
+                not account_id and
+                not event_id and
+                not organization_id and
+                not place_id and
+                not post_id):
+            return Response({
+                "message":  _("Neither Account, nor Event, nor Organization, "
+                              "nor Place, nor Post ID provided."),
+            }, status=status.HTTP_400_BAD_REQUEST)
+
+        if not rating:
+            return Response({
+                "message":  _("Rating is not provided."),
+            }, status=status.HTTP_400_BAD_REQUEST)
+
+        if not review_text:
+            return Response({
+                "message":  _("Review Text is not provided."),
+            }, status=status.HTTP_400_BAD_REQUEST)
+
+        # ---------------------------------------------------------------------
+        # --- Retrieve the Account.
+        # ---------------------------------------------------------------------
+        if account_id:
+            obj = get_object_or_None(User, id=account_id)
+            if not obj:
+                return Response({
+                    "message":  _("Member not found."),
+                }, status=status.HTTP_404_NOT_FOUND)
+
+            # -----------------------------------------------------------------
+            # --- Check, if the User has already rated the Account.
+            if obj.profile.is_rated_by_user(request.user):
+                return Response({
+                    "message":      _("You already rated the Member."),
+                }, status=status.HTTP_400_BAD_REQUEST)
+
+        # ---------------------------------------------------------------------
+        # --- Retrieve the Event.
+        # ---------------------------------------------------------------------
+        if event_id:
+            obj = get_object_or_None(Event, id=event_id)
+            if not obj:
+                return Response({
+                    "message":  _("Event not found."),
+                }, status=status.HTTP_404_NOT_FOUND)
+
+            # -----------------------------------------------------------------
+            # --- Check, if the User has already rated the Account.
+            if obj.is_rated_by_user(request.user):
+                return Response({
+                    "message":      _("You already rated the Event."),
+                }, status=status.HTTP_400_BAD_REQUEST)
+
+        # ---------------------------------------------------------------------
+        # --- Retrieve the Organization.
+        # ---------------------------------------------------------------------
+        if organization_id:
+            obj = get_object_or_None(Organization, id=organization_id)
+            if not obj:
+                return Response({
+                    "message":      _("Organization not found."),
+                }, status=status.HTTP_404_NOT_FOUND)
+
+            # -----------------------------------------------------------------
+            # --- Check, if the User has already rated the Organization.
+            if obj.is_rated_by_user(request.user):
+                return Response({
+                    "message":      _("You already rated the Organization."),
+                }, status=status.HTTP_400_BAD_REQUEST)
+
+        # ---------------------------------------------------------------------
+        # --- Retrieve the Place.
+        # ---------------------------------------------------------------------
+
+        # ---------------------------------------------------------------------
+        # --- Retrieve the Post.
+        # ---------------------------------------------------------------------
+        if post_id:
+            obj = get_object_or_None(Post, id=post_id)
+            if not obj:
+                return Response({
+                    "message":      _("Blog Post not found."),
+                }, status=status.HTTP_404_NOT_FOUND)
+
+            # -----------------------------------------------------------------
+            # --- Check, if the User has already rated the Post.
+            if obj.is_rated_by_user(request.user):
+                return Response({
+                    "message":      _("You already rated the Post."),
+                }, status=status.HTTP_400_BAD_REQUEST)
+
+        # ---------------------------------------------------------------------
+        # --- Create Rating.
+        # ---------------------------------------------------------------------
+        rating, created = Rating.objects.get_or_create(
+            author=request.user,
+            rating=int(rating),
+            review_text=review_text,
+            content_type=obj,
+            object_id=obj.id)
+        rating.save()
+
+        if author_rating:
+            rating, created = Rating.objects.get_or_create(
+                author=request.user,
+                rating=int(author_rating),
+                review_text=review_text,
+                content_type=obj.author,
+                object_id=obj.author.id)
+            rating.save()
+
+        return Response({
+            "message":      _("Successfully added the Rating."),
+        }, status=status.HTTP_200_OK)
+
+
+rating_list = RatingListViewSet.as_view()
+
+
+class RatingDetailsViewSet(APIView):
+    """Rating Details View Set."""
+
+    authentication_classes = (CsrfExemptSessionAuthentication, )
+    permission_classes = (IsAuthenticated, )
+    renderer_classes = (JSONRenderer, )
+    # serializer_class = RatingSerializer
+    # model = Rating
+
+    @log_default(my_logger=logger)
+    def delete(self, request, rating_id):
+        """DELETE: Rating delete.
+
+            Receive:
+
+                rating_id               :uint:
+
+            Return:
+
+                status                  200/400/404/500
+
+            Example Payload:
+
+                {
+                    "rating_id":        100,
+                }
+        """
+        # ---------------------------------------------------------------------
+        # --- Retrieve Data from the Request
+        # ---------------------------------------------------------------------
+
+        # ---------------------------------------------------------------------
+        # --- Verify Request.
+        # ---------------------------------------------------------------------
+        if not rating_id:
+            return Response({
+                "message":      _("No Rating ID provided."),
+            }, status=status.HTTP_400_BAD_REQUEST)
+
+        # ---------------------------------------------------------------------
+        # --- Retrieve the Rating
+        # ---------------------------------------------------------------------
+        rating = get_object_or_None(Rating, pk=rating_id)
+        if not rating:
+            return Response({
+                "message":      _("Rating not found."),
+            }, status=status.HTTP_404_NOT_FOUND)
+
+        if not request.user.is_staff:
+            return Response({
+                "message":      _("You don't have Permissions to perform the Action."),
+            }, status=status.HTTP_400_BAD_REQUEST)
+
+        rating.is_deleted = True
+        rating.save()
+
+        return Response({
+            "message":  _("Successfully removed the Rating."),
+        }, status=status.HTTP_200_OK)
+
+
+rating_details = RatingDetailsViewSet.as_view()
