+++ src/api/v1/collections/tests/test_collection_list.py
@@ -1,0 +1,3 @@
+"""
+(C) 2013-2025 Copycat Software, LLC. All Rights Reserved.
+"""
